finish all endpint for orders services

+++ eCommerceSolution.OrderService/BusinessLogicLayer/Services/OrdersService.cs
@@ -29,18 +29,37 @@ public class OrdersService(IOrdersRepository repository) : IOrdersService
 
     public async Task<OrderResponse?> AddOrder(OrderAddRequest orderAddRequest)
     {
+        if (orderAddRequest is null)
+        {
+            throw new ArgumentNullException(nameof(orderAddRequest), "OrderAddRequest cannot be null.");
+        }
+
         var added = await repository.AddOrder(orderAddRequest.ToEntity());
-        return added?.ToResponse();
+
+        // TO DO: Add logic for checking if userid exsits in user microservice, if not throw exception and return null
+
+        var reponse = added?.ToResponse();
+        return reponse;
     }
 
     public async Task<OrderResponse?> UpdateOrder(OrderUpdateRequest orderUpdateRequest)
     {
+        if (orderUpdateRequest is null)
+        {
+            throw new ArgumentNullException(nameof(orderUpdateRequest), "OrderUpdateRequest cannot be null.");
+        }
+
         var updated = await repository.UpdateOrder(orderUpdateRequest.ToEntity());
         return updated?.ToResponse();
     }
 
     public async Task<bool> DeleteOrder(Guid orderID)
     {
+        if (orderID == Guid.Empty)
+        {
+            throw new ArgumentNullException(nameof(orderID), "OrderID cannot be empty.");
+        }
+
         return await repository.DeleteOrder(orderID);
     }
 }
